THE A STAR IS DONE! (mostly lol)
The enemy moves only if the player moves and is within a certain distance of the player. as of right now each move the player does calculates the new node's distance and each enemies distance. which then sends a signal to each activated enemy's controller.  signaling to deecide on a move. if the player distance is 1 or less they attack. i just realized as i type this that i need to make them get one attack off while the player tries to escape. so if distance is <= 1 and player next node is 2 attack then the player gets a node inbetween and the enemy needs to get within 1 again in order to attack. also The attack parameter of being <= 1 isn't ideal. I want the attack only when the player attacks first. so
TLDR: Change enemy attack to only if the player attacks or if the player is on node away and then tries to escape!